Cherry-picked svg-android-2 patch #6
Patched to apply groups fill and stroke attributes to children.
Fixed doFill, which set black fill with no fill attribute.

This leaves out the implementation of the "use" tag, which to most
consumers of this library would be seen as a massive performance
regression.  Also the severe whitespace formatting errors were fixed in
the original patch.

Note that this patch does significantly improve the rendering accuracy
of the testInkscape unit test.  Still though, the top portion of the
oval shapes are not being rendered for unknown reasons.

Conflicts:
	src/main/java/com/larvalabs/svgandroid/SVGParser.java